Refactor out the Fire() method from ATank to UTankAimingComponent
This should remove all dependencies from ATank, and produce the final simplified architecture.
