+++ Source/BattleTank/Public/Tank.h
@@ -6,36 +6,12 @@
 #include "GameFramework/Pawn.h"
 #include "Tank.generated.h"
 
-class AProjectile;
-class UTankBarrel;
-
 UCLASS()
 class BATTLETANK_API ATank : public APawn
 {
 	GENERATED_BODY()
 
 public:
-	// Sets default values for this pawn's properties
+	// Sets default values for this component's properties
 	ATank();
-
-	UFUNCTION(BLueprintCallable)
-		void Fire();
-
-private:
-	UPROPERTY(EditDefaultsOnly, Category = "Setup")
-		TSubclassOf<AProjectile> ProjectileBlueprint;
-
-	// TODO Remove once firing is moved to AimingComponent
-	UPROPERTY(EditDefaultsOnly, Category = "Firing")
-		float LaunchSpeed = 4000;
-
-	UPROPERTY(EditDefaultsOnly, Category = "Firing")
-		float ReloadTimeInSeconds = 3;
-
-	// Local barrel for spawning projectiles
-	UTankBarrel* Barrel = nullptr;		// TODO Remove this reference to streamline
-
-	// Used to determine firing intervals, simulating reload time
-	double LastFireTime = 0;
-
 };
